// frontend/src/api/models/task-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * To do list API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TaskResponse
 */
export interface TaskResponse {
    /**
     * 
     * @type {number}
     * @memberof TaskResponse
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    'description'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof TaskResponse
     */
    'completed'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskResponse
     */
    'updatedAt'?: string | null;
}