// frontend/src/api/models/task-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * To do list API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TaskCreateRequest
 */
export interface TaskCreateRequest {
    /**
     * 
     * @type {string}
     * @memberof TaskCreateRequest
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof TaskCreateRequest
     */
    'description'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof TaskCreateRequest
     */
    'completed'?: boolean;
}