// frontend/src/api/models/task-update-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * To do list API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TaskUpdateRequest
 */
export interface TaskUpdateRequest {
    /**
     * 
     * @type {string}
     * @memberof TaskUpdateRequest
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TaskUpdateRequest
     */
    'description'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof TaskUpdateRequest
     */
    'completed'?: boolean | null;
}